Я забыл про коммиты есть только это

+++ Controllers/NoteController.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MyApi.Data;
@@ -10,23 +11,36 @@ namespace MyApi.Controllers;
 public class NoteController : ControllerBase
 {
     private readonly AppDbContext _db;
+    private readonly ILogger<NoteController> _logger;
     
-    public NoteController(AppDbContext db)
+    IAsyncEnumerable<Note> notes;  
+    
+    public NoteController(AppDbContext db, ILogger<NoteController> logger)  // ← добавить
     {
         _db = db;
+        _logger = logger;
     }
     
     [HttpPost]
-    public IActionResult CreateNote([FromBody] Note newNote)
+    [Authorize]
+    public async Task<IActionResult> CreateNote([FromBody] DTONote newNote)
     {
         if (newNote == null)
             return BadRequest("Нет данных");
-        
         if (string.IsNullOrEmpty(newNote.Title))
             return BadRequest("Имя обязательно");
-        newNote.CreatedAt = DateTime.Today;
-        _db.Notes.Add(newNote);
+
+        Note finNote = new Note();
+        var U = await _db.Users.FirstOrDefaultAsync(u => u.Email == User.Identity.Name);
+        if(U == null) return Unauthorized();
+        finNote.UserId = U.Id;
+        finNote.Title = newNote.Title;
+        finNote.Content = newNote.Content;
+        finNote.CreatedAt = DateTime.Now;
+        _db.Notes.Add(finNote);
         _db.SaveChanges();
+
+        _logger.LogInformation("Создание {0} c {1}", newNote.Title, finNote.Id);
 
         return Ok(new 
         { 
@@ -36,41 +50,89 @@ public class NoteController : ControllerBase
     }
     
     [HttpGet]
+    [Authorize]
     public async Task<IActionResult> GetAllNotes()
     {
-        var notes = await _db.Notes.ToListAsync();
-        return Ok(notes);
+        var U = await _db.Users.FirstOrDefaultAsync(u => u.Email == User.Identity.Name);
+        if(U == null) return Unauthorized();
+        if (User.IsInRole("Admin"))
+        {
+            var notes = await _db.Notes.Where(u => u.UserId == U.Id).ToListAsync();
+            return Ok(notes);
+        }
+        else
+        {
+            var notes = await _db.Notes.Where(u => u.UserId == U.Id).ToListAsync();
+            return Ok(notes);
+        }
     }
 
     [HttpGet("{Id}")]
-    public IActionResult GetNoteById(int Id)
+    [Authorize]
+    public async Task<IActionResult> GetNoteById(int Id)
     {
+        var U = await _db.Users.FirstOrDefaultAsync(u => u.Email == User.Identity.Name);
+        if(U == null) return Unauthorized();
+
         Note? nt = _db.Notes.Find(Id);
 
         if(nt == null)
             return NotFound("Note never exist");
 
+        if(nt.UserId != U.Id && !User.IsInRole("Admin")) return Unauthorized();
+
         return Ok(nt);
     }
 
     [HttpGet("search")]
+    [Authorize]
     public async Task<IActionResult> SearchNotes(
         [FromQuery] string? search,
         [FromQuery] DateTime? from,
-        [FromQuery] DateTime? to)
+        [FromQuery] DateTime? to,
+        [FromQuery] int? page,
+        [FromQuery] int? pageSize)
     {
+        var U = await _db.Users.FirstOrDefaultAsync(u => u.Email == User.Identity.Name);
+        if(U == null) return Unauthorized();
+
+
         List<Note> res = new List<Note>();
+        
+        if(page != 0 && pageSize != 0)
+        {
+            int actPage = page ?? 1;
+            int actSize = pageSize ?? 10;
+
+            res.AddRange(await _db.Notes.Where(u => u.UserId == U.Id).AsQueryable().Skip((actPage - 1)*actSize).Take(actSize).ToListAsync());
+        }
+        
         if (!string.IsNullOrWhiteSpace(search))
         {
-            var notes = _db.Notes.AsAsyncEnumerable();
-
+            if (User.IsInRole("Admin"))
+            {
+                var notes = _db.Notes.AsAsyncEnumerable();
+            }
+            else
+            {
+                var notes = _db.Notes.Where(u => u.UserId == U.Id).AsAsyncEnumerable();
+            }
+            
             var p = notes.Where(x => x.Title.ToUpper().Contains(search.ToUpper()) || x.Content.ToUpper().Contains(search.ToUpper()));
 
             res.AddRange(await p.ToListAsync());
         }
+        
         if(from != null)
         {
-            var notes = _db.Notes.AsAsyncEnumerable();
+            if (User.IsInRole("Admin"))
+            {
+                var notes = _db.Notes.AsAsyncEnumerable();
+            }
+            else
+            {
+                var notes = _db.Notes.Where(u => u.UserId == U.Id).AsAsyncEnumerable();
+            }
             if(notes == null) return NotFound("No any notes exist");
             await foreach(Note nt in notes)
             {
@@ -80,9 +142,17 @@ public class NoteController : ControllerBase
                 }
             }
         }
+        
         if(to != null)
         {
-            var notes = _db.Notes.AsAsyncEnumerable();
+            if (User.IsInRole("Admin"))
+            {
+                var notes = _db.Notes.AsAsyncEnumerable();
+            }
+            else
+            {
+                var notes = _db.Notes.Where(u => u.UserId == U.Id).AsAsyncEnumerable();
+            }
             if(notes == null) return NotFound("No any notes exist");
             await foreach(Note nt in notes)
             {
@@ -99,41 +169,55 @@ public class NoteController : ControllerBase
         return NotFound("No any notes found");
     }
 
-    [HttpPut]
-    public IActionResult PutNote([FromBody] Note newNote)
+    [HttpPut("{Id}")]
+    [Authorize]
+    public async Task<IActionResult> PutNote(int Id, [FromBody] DTONote newNote)
     {
+        var U = await _db.Users.FirstOrDefaultAsync(u => u.Email == User.Identity.Name);
+        if(U == null) return Unauthorized();
+
         if (newNote == null)
             return BadRequest("Нет данных");
         
         if (string.IsNullOrEmpty(newNote.Title))
             return BadRequest("Имя обязательно");
         
-        Note? nt = _db.Notes.Find(newNote.Id);
+        Note? nt = _db.Notes.Find(Id);
+
+        if(nt.UserId != U.Id && !User.IsInRole("Admin")) return Unauthorized();
 
         nt?.Title = newNote.Title;
         nt?.Content = newNote.Content;
         
         _db.SaveChanges();
 
+        _logger.LogInformation("Изменение {0} c {1}", newNote.Title, nt.Id);
+
         return Ok(new 
         { 
-            message = $"Заметка {newNote.Title} изменена",
-            note = newNote 
+            message = $"Заметка {newNote.Title} изменена"
         });
     }
 
     [HttpDelete("{Id}")]
-    public IActionResult DeleteNote(int Id)
+    [Authorize]
+    public async Task<IActionResult> DeleteNote(int Id)
     {
+        var U = await _db.Users.FirstOrDefaultAsync(u => u.Email == User.Identity.Name);
+        if(U == null) return Unauthorized();
         
         Note? nt = _db.Notes.Find(Id);
 
         if (nt == null)
             return  NotFound("Note never exist");
+
+        if(nt.UserId != U.Id && !User.IsInRole("Admin")) return Unauthorized();
         
         _db.Notes.Remove(nt);
         
         _db.SaveChanges();
+
+        _logger.LogInformation("Удаление {0} c {1}", nt.Title, Id);
 
         return Ok(new 
         {
